+++ aiootp/ciphers/decipher_streams.py
@@ -182,9 +182,7 @@ class AsyncDecipherStream(CipherStreamProperties, metaclass=AsyncInit):
         """
         return self.shmac.atest_next_block_id, self._buffer.append
 
-    async def _atest_timestamp(
-        self, queue: t.Callable[[bytes], None]
-    ) -> None:
+    async def _atest_timestamp(self, queue: t.Deque[bytes]) -> None:
         """
         Raises `TimestampExpired` if the timestamp prepended to the
         plaintext is older than the time-to-live specified by the
@@ -198,9 +196,7 @@ class AsyncDecipherStream(CipherStreamProperties, metaclass=AsyncInit):
             self._is_streaming = False
             raise error
 
-    async def _aremove_inner_header(
-        self, queue: t.Callable[[bytes], None]
-    ) -> None:
+    async def _aremove_inner_header(self, queue: t.Deque[bytes]) -> None:
         """
         Strips the inner header from the buffered plaintext in the queue
         in the cases where the inner header spans multiple blocks.
@@ -478,7 +474,7 @@ class DecipherStream(CipherStreamProperties):
         """
         return self.shmac.test_next_block_id, self._buffer.append
 
-    def _test_timestamp(self, queue: t.Callable[[bytes], None]) -> None:
+    def _test_timestamp(self, queue: t.Deque[bytes]) -> None:
         """
         Raises `TimestampExpired` if the timestamp prepended to the
         plaintext is older than the time-to-live specified by the
@@ -492,9 +488,7 @@ class DecipherStream(CipherStreamProperties):
             self._is_streaming = False
             raise error
 
-    def _remove_inner_header(
-        self, queue: t.Callable[[bytes], None]
-    ) -> None:
+    def _remove_inner_header(self, queue: t.Deque[bytes]) -> None:
         """
         Strips the inner header from the buffered plaintext in the queue
         in the cases where the inner header spans multiple blocks.